quorum update code

+++ quorumapi/views.py
@@ -3,17 +3,20 @@ from django.contrib.auth.models import User
 from django.core.exceptions import ValidationError
 from django.http import JsonResponse
 from django.shortcuts import get_object_or_404, redirect
+from django.http import HttpResponseRedirect
 from rest_framework import generics, permissions, viewsets
 from rest_framework.filters import OrderingFilter, SearchFilter
 from rest_framework.permissions import (IsAuthenticated,
                                         IsAuthenticatedOrReadOnly)
-
-from .models import Answer, Comment, Follow, Profile, Question
+from vote.views import VoteMixin
+from .models import (Answer, Comment, Follow, Profile, Question, QuestionVote,
+                    AnswerVote, CommentVote)
 from .paginators import UserPagination
 from .permissions import IsInstanceOwner
 from .serializers import (AnswerSerializer, CommentSerializer,
                           FollowerSerializer, ProfileSerializer,
-                          QuestionSerializer)
+                          QuestionSerializer, QuestionVoteSerializer,
+                          AnswerVoteSerializer, CommentVoteSerializer)
 
 
 class UserList(viewsets.ModelViewSet):
@@ -28,28 +31,29 @@ class UserList(viewsets.ModelViewSet):
         return serializer.save(user=self.request.user)
 
 
-@login_required
-def follow(request, pk):
+# @login_required
+def follow(request, pk): 
     user = get_object_or_404(User, pk=pk)
+    if request.user.is_authenticated: 
+        if user != request.user:
+            already_followed = Follow.objects.filter(
+                user=user, follower=request.user).first()
+            if not already_followed:
 
-    if user != request.user:
-        already_followed = Follow.objects.filter(
-            user=user, follower=request.user).first()
-        if not already_followed:
-
-            new_follower = Follow(user=user, follower=request.user)
-            new_follower.save()
-            follower_count = Follow.objects.filter(user=user).count()
-            return JsonResponse(
-                {'status': 'Following', 'count': follower_count})
-        else:
-            already_followed.delete()
-            follower_count = Follow.objects.filter(
-                user=user).count()
-            return JsonResponse(
-                {'status': 'Not following', 'count': follower_count})
-        return redirect('/')
-    raise ValidationError("One Cannot follow themselves")
+                new_follower = Follow(user=user, follower=request.user)
+                new_follower.save()
+                follower_count = Follow.objects.filter(user=user).count()
+                return JsonResponse(
+                    {'status': 'Following', 'count': follower_count})
+            else:
+                already_followed.delete()
+                follower_count = Follow.objects.filter(
+                    user=user).count()
+                return JsonResponse(
+                    {'status': 'Not following', 'count': follower_count})
+            return redirect('/')
+        raise ValidationError("One Cannot follow themselves")
+    return HttpResponseRedirect('/')
 
 
 class Following(generics.ListCreateAPIView):
@@ -70,6 +74,58 @@ class QuestionViewSet(viewsets.ModelViewSet):
     pagination_class = UserPagination
     search_fields = ['user__username','question']
     permission_classes = [IsAuthenticated, IsInstanceOwner, ]
+
+    def perform_create(self, serializer):
+        return serializer.save(user=self.request.user)
+    
+    def post(self, request):
+        import pdb;pdb.set_trace()
+        question = get_object_or_404(Question, slug=request.data.get('slug'))
+        if request.user not in question.favourite.all():
+            question.favourite.add(request.user)
+            return Response(
+                {'detail': 'User added to post'},
+                status=status.HTTP_200_OK)
+        return Response(
+            {'detail': self.bad_request_message},
+            status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request):
+        question = get_object_or_404(Question, slug=request.data.get('slug'))
+        if request.user in question.favourite.all():
+            question.favourite.remove(request.user)
+            return Response(
+                {'detail': 'User removed from post'},
+                status=status.HTTP_204_NO_CONTENT)
+        return Response(
+            {'detail': self.bad_request_message},
+            status=status.HTTP_400_BAD_REQUEST)
+
+class QuestionVoteViewSet(viewsets.ModelViewSet):
+    queryset = QuestionVote.objects.all()
+    serializer_class = QuestionVoteSerializer
+    pagination_class = UserPagination
+    permission_classes = [IsAuthenticated, IsInstanceOwner, ]
+
+    def perform_create(self, serializer):
+        return serializer.save(user=self.request.user)
+
+
+class AnswerVoteViewSet(viewsets.ModelViewSet):
+    queryset = AnswerVote.objects.all()
+    serializer_class = AnswerVoteSerializer
+    pagination_class = UserPagination
+    permission_classes = [IsAuthenticatedOrReadOnly, IsInstanceOwner, ]
+
+    def perform_create(self, serializer):
+        return serializer.save(user=self.request.user)
+
+
+class CommentVoteViewSet(viewsets.ModelViewSet):
+    queryset = CommentVote.objects.all()
+    serializer_class = CommentVoteSerializer
+    pagination_class = UserPagination
+    permission_classes = [IsAuthenticatedOrReadOnly, IsInstanceOwner, ]
 
     def perform_create(self, serializer):
         return serializer.save(user=self.request.user)
@@ -95,7 +151,7 @@ class Followers(generics.ListCreateAPIView):
         return Follow.objects.filter(user=user).exclude(follower=user)
 
 
-class CommentViewSet(viewsets.ModelViewSet):
+class CommentViewSet(viewsets.ModelViewSet,VoteMixin):
     queryset = Comment.objects.all()
     serializer_class = CommentSerializer
     pagination_class = UserPagination
